Fix some bugs.

+++ utils/eval_server.py
@@ -3,7 +3,7 @@
 # @Author: Haozhe Xie
 # @Date:   2020-08-08 17:16:07
 # @Last Modified by:   Haozhe Xie
-# @Last Modified time: 2020-08-08 22:15:43
+# @Last Modified time: 2020-08-08 23:12:21
 # @Email:  cshzxie@gmail.com
 
 import argparse
@@ -19,8 +19,7 @@ import sys
 
 def get_args_from_command_line():
     parser = argparse.ArgumentParser(description="The argument parser of the runner")
-    parser.add_argument("ckpt_dir",
-                        help="The path to folder that saves checkpoints")
+    parser.add_argument("ckpt_dir", help="The path to folder that saves checkpoints")
     args = parser.parse_args()
     return args
 
@@ -49,16 +48,16 @@ def get_next_checkpoints(tested_checkpoints, now_checkpoints):
 
 def get_jf_mean(process, checkpoint):
     try:
-        logs = process.stdout.read().split("\n")
+        logs = process.stdout.read().decode('utf-8').split("\n")
         for log in logs:
             if log.find('[Test Summary]') != -1:
                 last_comma = log.rfind(',')
                 last_quot_mark = log.rfind('\'')
                 return float(log[last_comma + 3:last_quot_mark])
     except Exception as ex:
-        logging.warn(ex)
+        logging.warning(ex)
 
-    logging.warn('Failed to parse test logs for checkpoint[Name=%s].' % checkpoint)
+    logging.warning('Failed to parse test logs for checkpoint[Name=%s].' % checkpoint)
     return 0
 
 
@@ -89,14 +88,7 @@ def main():
     logging.info("Listening new checkpoints at %s ..." % args.ckpt_dir)
 
     while True:
-        checkpoints = sorted([f for f in os.listdir(args.ckpt_dir) if f.startswith("ckpt-epoch")])
-        checkpoint = get_next_checkpoints(tested_checkpoints, checkpoints)
-
-        if checkpoint is None:
-            time.sleep(30)    # Detect new checkpoints every 30 seconds
-            continue
-
-        # Assign a free gpu to test the checkpoint
+        # Waiting for free GPUs & Processing test results
         while len(free_gpus) == 0:
             time.sleep(15)    # Detect free gpus every 15 seconds
             for rp in running_processes:
@@ -111,13 +103,22 @@ def main():
                     # Add the results to the TensorBoard
                     ckpt_epoch = int(rp["checkpoint"][len('ckpt-epoch-'):-len('.pth')])
                     test_writer.add_scalar('Metric/JF-Mean', jf_mean, ckpt_epoch)
-                    # Update the best JF Mean results
+                    # Update the best results
                     if jf_mean > best_jf_mean:
-                        os.remove(os.path.join(args.ckpt_dir, best_checkpoint))
+                        if best_checkpoint is not None:
+                            os.remove(os.path.join(args.ckpt_dir, best_checkpoint))
+
                         best_jf_mean = jf_mean
                         best_checkpoint = rp["checkpoint"]
                     else:
                         os.remove(os.path.join(args.ckpt_dir, rp["checkpoint"]))
+
+        # Waiting for new checkpoints
+        checkpoints = sorted([f for f in os.listdir(args.ckpt_dir) if f.startswith("ckpt-epoch")])
+        checkpoint = get_next_checkpoints(tested_checkpoints, checkpoints)
+        if checkpoint is None:
+            time.sleep(30)    # Detect new checkpoints every 30 seconds
+            continue
 
         assigned_gpu = free_gpus.pop(0)
         logging.info("Assign GPU[ID=%s] for the checkpoint[Name=%s]." % (assigned_gpu, checkpoint))
